% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CountImpute2Subjects}
\alias{CountImpute2Subjects}
\title{Function to count the subject in an Impute 2 file}
\usage{
CountImpute2Subjects(i2file, dataStart, format)
}
\arguments{
\item{i2file}{Name of impute 2 file}

\item{dataStart}{Column where the genetic data starts. This is also the column
where the subject IDs start.}

\item{format}{Indicator if family IDs are included in the header}
}
\value{
List containing status and the number of subjects found
}
\description{
Function to count the subject in an Impute 2 file. This
requires that the file not have a header. No error checking
of the genetic values are done.
}
