% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindImpute2SNPData}
\alias{FindImpute2SNPData}
\title{Function to find the SNP data columns from an Impute 2 file}
\usage{
FindImpute2SNPData(i2file, dataStart)
}
\arguments{
\item{i2file}{Name of impute 2 file}

\item{dataStart}{Column where the genetic data starts. Only the columns before
the genetic data are checked for SNP data}
}
\value{
List containing status and an
Integer vector of length 5 will column numbers for
SNP name, chromosome, location in BP, reference allele
and alternate allele. These columns must have column
name 'SNP', 'CHR', 'BP', 'A1', 'A2', respectively
}
\description{
Function to find the SNP data columns from an Impute 2 file.
Checks that the SNP name column exists and that there are no
duplicate columns.
}
