% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute2.R
\name{GetI2Info}
\alias{GetI2Info}
\title{Function to get info about Impute 2 file}
\usage{
GetI2Info(i2file, header = TRUE, snpCol, startCol, format = 2L,
  usesFID = TRUE, sep = "\\t")
}
\arguments{
\item{i2file}{Impute 2 file name}

\item{header}{Does the Impute 2 file have a header?}

\item{snpCol}{Must be provided if there is no header.
Vector of column numbers for SNP name, chromosome, location in BP,
reference allele, and alternate allele respectively. All must be provided
a value of 0 indicates the value isn't in the file. SNP name column is required.
If provided and there is a header, column names in header are ignored.
If not provided, column names in header are used to find appropriate columns
In this case the column names must be 'SNP', 'CHR', 'BP', 'A1' and 'A2'.}

\item{format}{Number of genetic probabilities in file, must be 1, 2 or 3
1 indicates that only the dosage is in the file}

\item{usesFID}{Does the impute file contain both the family and subject ID? Not used if
there is not header}

\item{sep}{Separator used in file. Currently only space and tabs are allowed}

\item{startcol}{Column number where dosages or genetic probabilities begin. Must be provided}
}
\description{
Function to get info about Impute 2 file
}
