% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute2toBD.R
\docType{package}
\name{Impute2toBD}
\alias{Impute2toBD}
\alias{Impute2toBD-package}
\title{Impute2toBD: A package for creating binary dosage files from Impute 2 files}
\description{
Impute2toBD: A package for creating binary dosage files from Impute 2 files
}
\section{Functions}{

Currently none
}

