% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Impute2toBD_C}
\alias{Impute2toBD_C}
\title{Function to convert and Impute 2 file to binary dosage format}
\usage{
Impute2toBD_C(impute2File, header, snpInfoCols, dataStartCol, binaryDosageFile)
}
\arguments{
\item{impute2File}{Name of Impute 2 file}

\item{header}{Does the impute 2 file have a header}

\item{snpInfoCols}{Column numbers for chromosome, SNP name, location in BP, allele 1 and allele 2, respecitively}

\item{dataStartCol}{Column where dosage/genetic probabilities start}

\item{binaryDosageFile}{Name of output file}
}
\value{
String reporting status
}
\description{
Function to convert and Impute 2 file to binary dosage format
}
