% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadImpute2Subjects}
\alias{ReadImpute2Subjects}
\title{Function to read the subject IDs from an Impute 2 file}
\usage{
ReadImpute2Subjects(i2file, dataStart, usesFID)
}
\arguments{
\item{i2file}{Name of impute 2 file}

\item{dataStart}{Column where the genetic data starts. This is also the column
where the subject IDs start.}

\item{usesFID}{Indicator if family IDs are included in the header}
}
\value{
List containing status and a
data frame with the subject and family IDs. If no family ID is
in the file these values are set to blank, "", not missing.
}
\description{
Function to read the subject IDs from an Impute 2 file. This
requires that the file have a header
}
